services/xrpl/requirements.txt
fastapi==0.95.1
requests==2.28.1
ptvsd==4.3.2 # Required for debugging.
gunicorn== 20.1.0
uvicorn==0.22.0
websockets==10.4
xrpl-py==1.8.0
xumm-sdk-py==1.0.3
asyncio==3.4.3
nest_asyncio==1.5.6
google-cloud-storage
google-cloud-secret-manager
google-cloud-functions
google-cloud-firestore==2.11.1
Jinja2==3.1.2
pydantic